chore: consolidate conftest files for losses
